# Conflicts:
#	DigitalFUHubApi/Controllers/UsersController.cs

+++ DataAccess/Repositories/FeedbackRepository.cs
@@ -15,7 +15,7 @@ namespace DataAccess.Repositories
 		public Order? GetFeedbackDetail(long orderId, long userId)
 		=> FeedbackDAO.Instance.GetFeedbackDetail(orderId, userId);
 
-		public int AddFeedbackOrder(long userId, long orderId, long orderDetailId, string content, int rate, List<string> urlImages)
+		public long AddFeedbackOrder(long userId, long orderId, long orderDetailId, string content, int rate, List<string> urlImages)
 		=> FeedbackDAO.Instance.AddFeedbackOrder(userId, orderId, orderDetailId, content, rate, urlImages);
 
 		public List<FeedbackResponseDTO> GetFeedbacks(long productId)
